net: openthread: Add openthread tcat multiradio config.
As per BHC-750 in TCAT case for now only allowed devices are the ones
that have either dual radio or multiplexed radio capabilities. So the 'M' bit
in BLE advertisement should be always set to 1. In future specifications
devices without multiradio capabilities will be allowed this is the reason
why new Kconfig was introduced.

Signed-off-by: Przemyslaw Bida <przemyslaw.bida@nordicsemi.no>